move bulk of context menu actions to header menu
